+++ src/main/java/Coursework/WrongRequestException.java
@@ -1,0 +1,7 @@
+package Coursework;
+
+public class WrongRequestException extends RuntimeException{
+    public WrongRequestException(String message) {
+        super(message);
+    }
+}
